ast_results/andstatus_andstatus/app/src/main/java/org/andstatus/app/account/AccountSettingsActivity.java
// isComment
package org.andstatus.app.account;

import android.accounts.AccountManager;
import android.app.ProgressDialog;
import android.content.Intent;
import android.net.Uri;
import android.os.Bundle;
import android.support.annotation.IdRes;
import android.support.v4.app.Fragment;
import android.text.Editable;
import android.text.TextUtils;
import android.text.TextWatcher;
import android.view.KeyEvent;
import android.view.Menu;
import android.view.MenuItem;
import android.view.View;
import android.view.WindowManager;
import android.widget.EditText;
import android.widget.TextView;
import android.widget.Toast;
import com.github.scribejava.core.model.OAuth2AccessToken;
import com.github.scribejava.core.oauth.OAuth20Service;
import org.andstatus.app.ActivityRequestCode;
import org.andstatus.app.IntentExtra;
import org.andstatus.app.MyActivity;
import org.andstatus.app.R;
import org.andstatus.app.account.MyAccount.CredentialsVerificationStatus;
import org.andstatus.app.context.MyContext;
import org.andstatus.app.context.MyContextHolder;
import org.andstatus.app.context.MyPreferences;
import org.andstatus.app.context.MySettingsActivity;
import org.andstatus.app.net.http.ConnectionException;
import org.andstatus.app.net.http.HttpConnection;
import org.andstatus.app.origin.Origin;
import org.andstatus.app.origin.OriginType;
import org.andstatus.app.origin.PersistentOriginList;
import org.andstatus.app.os.AsyncTaskLauncher;
import org.andstatus.app.os.MyAsyncTask;
import org.andstatus.app.service.CommandData;
import org.andstatus.app.service.CommandEnum;
import org.andstatus.app.service.MyServiceManager;
import org.andstatus.app.service.MyServiceState;
import org.andstatus.app.timeline.TimelineActivity;
import org.andstatus.app.timeline.meta.Timeline;
import org.andstatus.app.timeline.meta.TimelineType;
import org.andstatus.app.util.DialogFactory;
import org.andstatus.app.util.I18n;
import org.andstatus.app.util.MyCheckBox;
import org.andstatus.app.util.MyLog;
import org.andstatus.app.util.MyUrlSpan;
import org.andstatus.app.util.RelativeTime;
import org.andstatus.app.util.SharedPreferencesUtil;
import org.andstatus.app.util.StringUtils;
import org.andstatus.app.util.TriState;
import org.andstatus.app.util.ViewUtils;
import org.andstatus.app.view.EnumSelector;
import org.json.JSONException;
import org.json.JSONObject;
import java.util.List;
import oauth.signpost.OAuth;
import oauth.signpost.OAuthConsumer;
import oauth.signpost.exception.OAuthCommunicationException;
import oauth.signpost.exception.OAuthExpectationFailedException;
import oauth.signpost.exception.OAuthMessageSignerException;
import oauth.signpost.exception.OAuthNotAuthorizedException;

/**
 * isComment
 */
public class isClassOrIsInterface extends MyActivity {

    private static final String isVariable = AccountSettingsActivity.class.isMethod();

    private enum ResultStatus {

        NONE, ACCOUNT_VALID, ACCOUNT_INVALID, CONNECTION_EXCEPTION, CREDENTIALS_OF_OTHER_ACCOUNT
    }

    private static class isClassOrIsInterface {

        final ResultStatus isVariable;

        CharSequence isVariable = "isStringConstant";

        isConstructor(ResultStatus isParameter) {
            this(isNameExpr, "isStringConstant");
        }

        isConstructor(ResultStatus isParameter, CharSequence isParameter) {
            this.isFieldAccessExpr = isNameExpr;
            this.isFieldAccessExpr = isNameExpr;
        }

        boolean isMethod() {
            return isNameExpr == isNameExpr.isFieldAccessExpr;
        }
    }

    private static final String isVariable = "isStringConstant";

    private static final String isVariable = "isStringConstant";

    private enum ActivityOnFinish {

        NONE, HOME, OUR_DEFAULT_SCREEN
    }

    private volatile ActivityOnFinish isVariable = isNameExpr.isFieldAccessExpr;

    private volatile boolean isVariable = true;

    private StateOfAccountChangeProcess isVariable = null;

    private StringBuilder isVariable = new StringBuilder();

    private boolean isVariable = true;

    @Override
    protected void isMethod(Bundle isParameter) {
        isNameExpr = true;
        isNameExpr = isNameExpr.isFieldAccessExpr.isFieldAccessExpr;
        super.isMethod(isNameExpr);
        if (isNameExpr.isMethod(this)) {
            return;
        }
        if (isNameExpr == null) {
            isMethod(AccountSettingsFragment.class);
        }
        isMethod().isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
    }

    private boolean isMethod(@IdRes int isParameter) {
        return !isMethod(isNameExpr);
    }

    private boolean isMethod(@IdRes int isParameter) {
        View isVariable = isMethod(isNameExpr);
        return isNameExpr != null && isNameExpr.isMethod() == isNameExpr.isFieldAccessExpr;
    }

    private View isMethod(@IdRes int isParameter) {
        Fragment isVariable = isMethod().isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
        if (isNameExpr != null) {
            View isVariable = isNameExpr.isMethod();
            if (isNameExpr != null) {
                return isNameExpr.isMethod(isNameExpr);
            }
        }
        return null;
    }

    @Override
    protected void isMethod(Intent isParameter) {
        super.isMethod(isNameExpr);
        isMethod(isNameExpr);
        isMethod(isNameExpr, "isStringConstant");
    }

    /**
     * isComment
     */
    protected void isMethod(Intent isParameter, String isParameter) {
        String isVariable = "isStringConstant";
        if (isNameExpr == null) {
            isNameExpr = isNameExpr.isMethod();
            isNameExpr += (isNameExpr.isFieldAccessExpr ? "isStringConstant" : "isStringConstant");
        } else {
            isNameExpr += "isStringConstant" + (isNameExpr.isFieldAccessExpr ? "isStringConstant" : "isStringConstant");
        }
        StateOfAccountChangeProcess isVariable = isNameExpr.isMethod(isNameExpr);
        if (isNameExpr.isFieldAccessExpr || isNameExpr.isFieldAccessExpr) {
            isNameExpr += "isStringConstant";
            isNameExpr = isNameExpr;
            if (isNameExpr.isFieldAccessExpr) {
                isNameExpr.isMethod(isNameExpr.isFieldAccessExpr, OriginType.class).isMethod(this);
            } else if (isNameExpr.isFieldAccessExpr) {
                isNameExpr.isMethod(this, isNameExpr.isFieldAccessExpr, isIntegerConstant);
                isNameExpr += "isStringConstant";
            }
            isNameExpr += "isStringConstant" + isNameExpr.isMethod() + "isStringConstant";
            isMethod();
        }
        if (isNameExpr.isFieldAccessExpr != null) {
            isNameExpr += "isStringConstant";
        }
        isNameExpr.isMethod(this, "isStringConstant" + isNameExpr + "isStringConstant" + isNameExpr + "isStringConstant" + isNameExpr.isMethod(isIntegerConstant));
    }

    @Override
    public void isMethod(int isParameter, int isParameter, Intent isParameter) {
        switch(isNameExpr.isMethod(isNameExpr)) {
            case isNameExpr:
                isMethod(isNameExpr, isNameExpr);
                break;
            case isNameExpr:
                isMethod(isNameExpr, isNameExpr);
                break;
            case isNameExpr:
                isMethod(isNameExpr, isNameExpr);
                break;
            default:
                super.isMethod(isNameExpr, isNameExpr, isNameExpr);
                break;
        }
    }

    private void isMethod(int isParameter, Intent isParameter) {
        if (isNameExpr == isNameExpr) {
            isNameExpr.isFieldAccessExpr = isNameExpr.isFieldAccessExpr.isMethod(isNameExpr.isMethod(), isNameExpr.isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr), isNameExpr.isFieldAccessExpr);
            if (!isNameExpr.isFieldAccessExpr.isMethod()) {
                isNameExpr = true;
            }
        } else {
            isNameExpr = true;
        }
        if (!isNameExpr) {
            isNameExpr.isMethod(this, "isStringConstant");
            isNameExpr.isFieldAccessExpr.isFieldAccessExpr.isMethod().isMethod(isNameExpr.isFieldAccessExpr.isMethod());
            isNameExpr.isMethod(isNameExpr.isFieldAccessExpr);
            isMethod();
        } else {
            isNameExpr.isMethod(this, "isStringConstant");
            isMethod();
        }
    }

    private void isMethod(int isParameter, Intent isParameter) {
        OriginType isVariable = isNameExpr.isFieldAccessExpr;
        if (isNameExpr == isNameExpr) {
            isNameExpr = isNameExpr.isMethod(isNameExpr.isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr));
            if (isNameExpr.isMethod()) {
                List<Origin> isVariable = isNameExpr.isFieldAccessExpr.isFieldAccessExpr.isMethod().isMethod(isNameExpr);
                switch(isNameExpr.isMethod()) {
                    case isIntegerConstant:
                        isNameExpr = isNameExpr.isFieldAccessExpr;
                        break;
                    case isIntegerConstant:
                        isMethod(isNameExpr.isMethod(isIntegerConstant));
                        break;
                    default:
                        isMethod(isNameExpr);
                        break;
                }
            }
        }
        if (!isNameExpr.isMethod()) {
            isMethod();
        }
    }

    protected boolean isMethod(OriginType isParameter) {
        Intent isVariable = new Intent(isNameExpr.this, PersistentOriginList.class);
        isNameExpr.isMethod(isNameExpr.isFieldAccessExpr);
        isNameExpr.isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr, isNameExpr.isMethod());
        isMethod(isNameExpr, isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
        return true;
    }

    private void isMethod(int isParameter, Intent isParameter) {
        Origin isVariable = isNameExpr.isFieldAccessExpr;
        if (isNameExpr == isNameExpr) {
            isNameExpr = isNameExpr.isFieldAccessExpr.isFieldAccessExpr.isMethod().isMethod(isNameExpr.isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr));
            if (isNameExpr.isMethod()) {
                isMethod(isNameExpr);
            }
        }
        if (!isNameExpr.isMethod()) {
            isMethod();
        }
    }

    private void isMethod(Origin isParameter) {
        if (!isNameExpr.isMethod().isMethod().isMethod(isNameExpr)) {
            // isComment
            isNameExpr.isFieldAccessExpr = isNameExpr.isFieldAccessExpr.isMethod(isNameExpr.isMethod(), isNameExpr.isMethod(isNameExpr, isNameExpr.isMethod().isMethod()).isMethod(), isNameExpr.isMethod(isNameExpr.isMethod().isMethod()));
            isMethod();
            isMethod();
        }
    }

    private void isMethod() {
        if (isNameExpr.isMethod().isMethod(isNameExpr.isFieldAccessExpr) && isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr) && isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr) && isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr)) {
            View isVariable = isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
            if (isNameExpr != null)
                isNameExpr.isMethod();
        }
    }

    @Override
    public boolean isMethod(Menu isParameter) {
        isMethod().isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr, isNameExpr);
        return super.isMethod(isNameExpr);
    }

    @Override
    public boolean isMethod(Menu isParameter) {
        MenuItem isVariable = isNameExpr.isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
        if (isNameExpr != null) {
            final boolean isVariable = isNameExpr != null && isNameExpr.isFieldAccessExpr != null && isNameExpr.isFieldAccessExpr.isMethod();
            isNameExpr.isMethod(isNameExpr);
            isNameExpr.isMethod(isNameExpr);
        }
        return super.isMethod(isNameExpr);
    }

    @Override
    public boolean isMethod(MenuItem isParameter) {
        switch(isNameExpr.isMethod()) {
            case isNameExpr.isFieldAccessExpr.isFieldAccessExpr.isFieldAccessExpr:
                isMethod();
                return true;
            case isNameExpr.isFieldAccessExpr.isFieldAccessExpr:
                isMethod();
                break;
            case isNameExpr.isFieldAccessExpr.isFieldAccessExpr:
                isNameExpr.isMethod(isMethod().isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr), isNameExpr.isFieldAccessExpr.isFieldAccessExpr, isNameExpr.isFieldAccessExpr.isFieldAccessExpr, isNameExpr.isFieldAccessExpr);
                break;
            default:
                break;
        }
        return super.isMethod(isNameExpr);
    }

    private void isMethod() {
        isMethod();
        isMethod(new Intent(this, MySettingsActivity.class));
    }

    private void isMethod() {
        if (isMethod().isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr) == null) {
            isNameExpr.isMethod(this, "isStringConstant");
            return;
        }
        isMethod();
        isMethod();
        isMethod();
        isMethod();
        isMethod();
        isMethod();
        isMethod();
        isNameExpr.isMethod(this, isNameExpr.isFieldAccessExpr.isFieldAccessExpr, isNameExpr.isFieldAccessExpr.isMethod());
        isMethod();
        isMethod();
        isMethod();
        isMethod();
        isMethod();
        isMethod();
    }

    private void isMethod() {
        MyAccount isVariable = isNameExpr.isMethod();
        String isVariable = isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr).isMethod();
        if (isNameExpr.isMethod()) {
            isNameExpr += "isStringConstant" + isNameExpr.isMethod();
        }
        isMethod(isNameExpr);
    }

    private void isMethod() {
        isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr, isNameExpr.isFieldAccessExpr.isFieldAccessExpr, isNameExpr.isMethod() > isIntegerConstant);
        isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr, isNameExpr, isNameExpr.isMethod() > isIntegerConstant);
    }

    private void isMethod() {
        MyAccount isVariable = isNameExpr.isMethod();
        TextView isVariable = (TextView) isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
        if (isNameExpr != null) {
            isNameExpr.isMethod(this.isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr).isMethod().isMethod("isStringConstant", isNameExpr.isMethod().isMethod()).isMethod("isStringConstant", isNameExpr.isMethod().isMethod().isMethod()));
        }
    }

    private void isMethod() {
        MyAccount isVariable = isNameExpr.isMethod();
        isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr, isNameExpr.isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr), isNameExpr.isFieldAccessExpr.isMethod() || isNameExpr.isMethod());
        EditText isVariable = (EditText) isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
        if (isNameExpr != null) {
            if (isNameExpr.isFieldAccessExpr.isMethod() || !isNameExpr.isMethod()) {
                isNameExpr.isMethod(isNameExpr.isFieldAccessExpr);
            } else {
                isNameExpr.isMethod(isNameExpr.isFieldAccessExpr);
                isNameExpr.isMethod(isNameExpr.isMethod(isMethod(isNameExpr.isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr)).isMethod(), isNameExpr.isMethod().isMethod(), isNameExpr.isMethod().isMethod().isFieldAccessExpr));
                isNameExpr.isMethod(isNameExpr);
                if (isNameExpr.isMethod().isMethod() == isIntegerConstant) {
                    isNameExpr.isMethod();
                }
            }
            if (isNameExpr.isMethod().isMethod(isNameExpr.isMethod().isMethod()) != isIntegerConstant) {
                isNameExpr.isMethod(isNameExpr.isMethod());
            }
            isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr, isNameExpr.isMethod(), isNameExpr.isFieldAccessExpr.isMethod());
        }
    }

    private TextWatcher isVariable = new TextWatcher() {

        @Override
        public void isMethod(CharSequence isParameter, int isParameter, int isParameter, int isParameter) {
        // isComment
        }

        @Override
        public void isMethod(CharSequence isParameter, int isParameter, int isParameter, int isParameter) {
        // isComment
        }

        @Override
        public void isMethod(Editable isParameter) {
            isMethod();
        }
    };

    private void isMethod() {
        MyAccount isVariable = isNameExpr.isMethod();
        boolean isVariable = isNameExpr.isMethod().isMethod() && !isNameExpr.isMethod();
        StringBuilder isVariable = new StringBuilder();
        if (isNameExpr) {
            isNameExpr.isMethod(this.isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr));
            if (isNameExpr.isMethod(isNameExpr.isMethod())) {
                isNameExpr.isMethod("isStringConstant" + this.isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr) + "isStringConstant");
            }
        }
        isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr, isNameExpr.isMethod(), isNameExpr);
        EditText isVariable = (EditText) isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
        if (isNameExpr != null) {
            if (isNameExpr.isMethod().isMethod(isNameExpr.isMethod().isMethod()) != isIntegerConstant) {
                isNameExpr.isMethod(isNameExpr.isMethod());
            }
            isNameExpr.isMethod(isNameExpr ? isNameExpr.isFieldAccessExpr : isNameExpr.isFieldAccessExpr);
            isNameExpr.isMethod(!isNameExpr.isMethod());
            isNameExpr.isMethod(isNameExpr);
        }
    }

    private void isMethod() {
        MyAccount isVariable = isNameExpr.isMethod();
        StringBuilder isVariable = null;
        if (isNameExpr.isFieldAccessExpr.isMethod()) {
            switch(isNameExpr.isMethod()) {
                case isNameExpr:
                    isNameExpr = new StringBuilder(this.isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr));
                    break;
                default:
                    if (isNameExpr.isFieldAccessExpr.isMethod()) {
                        isNameExpr = new StringBuilder(this.isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr));
                    } else {
                        if (isNameExpr.isMethod()) {
                            isNameExpr = new StringBuilder(this.isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr));
                        } else {
                            isNameExpr = new StringBuilder(this.isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr));
                        }
                    }
                    break;
            }
        }
        TextView isVariable = (TextView) isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
        if (isNameExpr != null) {
            isNameExpr.isMethod(isNameExpr);
        }
    }

    private void isMethod() {
        TextView isVariable = isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr, null, !isNameExpr.isFieldAccessExpr.isMethod());
        if (isNameExpr != null && isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr)) {
            isNameExpr.isMethod(isParameter -> {
                isMethod();
                isMethod();
                isMethod();
                MyAccount isVariable = isNameExpr.isMethod();
                CharSequence isVariable = "isStringConstant";
                boolean isVariable = !isNameExpr.isMethod() || isNameExpr.isMethod();
                if (isNameExpr) {
                    if (!isNameExpr.isMethod() && !isNameExpr.isMethod()) {
                        isNameExpr = true;
                        isNameExpr = isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
                    }
                } else {
                    isNameExpr = isMethod(isNameExpr.isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr));
                }
                if (isNameExpr) {
                    isMethod(true);
                } else {
                    isMethod(isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr) + "isStringConstant" + isNameExpr);
                }
            });
        }
    }

    private void isMethod() {
        android.webkit.CookieManager isVariable = isNameExpr.isFieldAccessExpr.isFieldAccessExpr.isMethod();
        isNameExpr.isMethod();
        if (isNameExpr.isMethod() > isIntegerConstant) {
            isNameExpr.isMethod(isIntegerConstant);
            isMethod();
        }
    }

    private void isMethod() {
        TextView isVariable = isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr, isNameExpr.isFieldAccessExpr.isFieldAccessExpr, isNameExpr.isFieldAccessExpr.isMethod());
        if (isNameExpr != null) {
            isNameExpr.isMethod(isParameter -> {
                isMethod();
                isNameExpr = isNameExpr.isFieldAccessExpr;
                isMethod();
            });
        }
    }

    private void isMethod() {
        TextView isVariable = isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr, isNameExpr.isMethod().isMethod() ? isNameExpr.isFieldAccessExpr.isFieldAccessExpr : isNameExpr.isFieldAccessExpr.isFieldAccessExpr, isNameExpr.isFieldAccessExpr.isMethod());
        if (isNameExpr != null) {
            isNameExpr.isMethod(isParameter -> {
                isMethod();
                isMethod();
                isMethod();
                isMethod(true);
            });
        }
    }

    private void isMethod() {
        boolean isVariable = isNameExpr.isMethod().isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr.isMethod().isMethod());
        View isVariable = isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
        if (isNameExpr != null) {
            isNameExpr.isMethod(isNameExpr ? isNameExpr.isFieldAccessExpr : isNameExpr.isFieldAccessExpr);
        }
    }

    private void isMethod() {
        isNameExpr.isMethod(isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr), isNameExpr.isFieldAccessExpr.isMethod().isMethod(), (isParameter, isParameter) -> isNameExpr.isFieldAccessExpr.isMethod(isNameExpr));
    }

    private void isMethod() {
        TextView isVariable = (TextView) isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
        EditText isVariable = (EditText) isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
        if (isNameExpr != null && isNameExpr != null) {
            String isVariable = isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr).isMethod() + "isStringConstant" + isNameExpr.isMethod(this, isNameExpr.isMethod(isNameExpr.isMethod()), isNameExpr.isFieldAccessExpr.isFieldAccessExpr, isNameExpr.isFieldAccessExpr.isFieldAccessExpr, isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
            isNameExpr.isMethod(isNameExpr);
            String isVariable = isNameExpr.isFieldAccessExpr.isMethod().isMethod() <= isIntegerConstant ? "isStringConstant" : isNameExpr.isMethod(isNameExpr.isFieldAccessExpr.isMethod().isMethod() / isIntegerConstant);
            isNameExpr.isMethod(isNameExpr);
            isNameExpr.isMethod(isNameExpr);
            isNameExpr.isMethod(new TextWatcher() {

                @Override
                public void isMethod(CharSequence isParameter, int isParameter, int isParameter, int isParameter) {
                // isComment
                }

                @Override
                public void isMethod(CharSequence isParameter, int isParameter, int isParameter, int isParameter) {
                // isComment
                }

                @Override
                public void isMethod(Editable isParameter) {
                    long isVariable = isNameExpr.isMethod(isNameExpr.isMethod());
                    isNameExpr.isFieldAccessExpr.isMethod(isNameExpr > isIntegerConstant ? isNameExpr * isIntegerConstant : isIntegerConstant);
                }
            });
        }
    }

    private void isMethod() {
        long isVariable = isNameExpr.isMethod().isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
        isNameExpr.isMethod((TextView) isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr), isNameExpr == isIntegerConstant ? isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr).isMethod() : isNameExpr.isMethod(this, isNameExpr), true, true);
    }

    private TextView isMethod(int isParameter, int isParameter, boolean isParameter) {
        return isMethod(isNameExpr, isNameExpr == isIntegerConstant ? null : isMethod(isNameExpr), isNameExpr);
    }

    private TextView isMethod(int isParameter, CharSequence isParameter, boolean isParameter) {
        TextView isVariable = (TextView) isMethod(isNameExpr);
        if (isNameExpr != null) {
            if (!isNameExpr.isMethod(isNameExpr)) {
                isNameExpr.isMethod(isNameExpr);
            }
            isNameExpr.isMethod(isNameExpr ? isNameExpr.isFieldAccessExpr : isNameExpr.isFieldAccessExpr);
        }
        return isNameExpr;
    }

    @Override
    protected void isMethod() {
        super.isMethod();
        isNameExpr.isMethod().isMethod(true);
        if (isNameExpr.isMethod(this)) {
            return;
        }
        isNameExpr.isMethod();
        isNameExpr.isMethod();
        isMethod();
        Uri isVariable = isMethod().isMethod();
        if (isNameExpr != null) {
            if (isNameExpr.isMethod(isNameExpr, isNameExpr.isFieldAccessExpr)) {
                isNameExpr.isMethod(isNameExpr, "isStringConstant" + isNameExpr.isMethod());
            }
            if (isNameExpr.isFieldAccessExpr.isMethod().isMethod(isNameExpr.isMethod())) {
                // isComment
                isMethod().isMethod(null);
                // isComment
                // isComment
                new AsyncTaskLauncher<Uri>().isMethod(this, true, new OAuthAcquireAccessTokenTask(), isNameExpr);
                isNameExpr = isNameExpr.isFieldAccessExpr;
            }
        }
        isNameExpr = true;
    }

    /**
     * isComment
     */
    private void isMethod(boolean isParameter) {
        MyAccount isVariable = isNameExpr.isMethod();
        if (isNameExpr || isNameExpr.isMethod() == isNameExpr.isFieldAccessExpr) {
            isNameExpr.isMethod();
            MyServiceState isVariable = isNameExpr.isMethod();
            if (isNameExpr != isNameExpr.isFieldAccessExpr) {
                isNameExpr.isMethod();
                if (isNameExpr != isNameExpr.isFieldAccessExpr) {
                    isMethod(isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr) + "isStringConstant" + isNameExpr + "isStringConstant");
                    return;
                }
            }
            if (isNameExpr.isMethod()) {
                // isComment
                // isComment
                isNameExpr.isMethod(this, true, new VerifyCredentialsTask());
            } else {
                if (isNameExpr.isMethod() && isNameExpr) {
                    // isComment
                    if (!isNameExpr.isMethod()) {
                        isNameExpr.isMethod(this, true, new OAuthRegisterClientTask());
                    } else {
                        isNameExpr.isMethod(this, true, new OAuthAcquireRequestTokenTask());
                        isNameExpr = isNameExpr.isFieldAccessExpr;
                    }
                }
            }
        }
    }

    private void isMethod() {
        if (!isNameExpr.isFieldAccessExpr.isMethod()) {
            EditText isVariable = (EditText) isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
            if (isNameExpr != null) {
                String isVariable = isNameExpr.isMethod().isMethod();
                if (isNameExpr.isMethod(isNameExpr.isMethod().isMethod()) != isIntegerConstant) {
                    boolean isVariable = isNameExpr.isMethod().isMethod();
                    isNameExpr.isFieldAccessExpr = isNameExpr.isFieldAccessExpr.isMethod(isNameExpr.isMethod(), isNameExpr.isMethod(isNameExpr.isMethod().isMethod(), isNameExpr).isMethod(), isNameExpr.isMethod(isNameExpr));
                }
            }
        }
        EditText isVariable = (EditText) isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
        if (isNameExpr != null && isNameExpr.isMethod().isMethod().isMethod(isNameExpr.isMethod().isMethod()) != isIntegerConstant) {
            isNameExpr.isFieldAccessExpr.isMethod(isNameExpr.isMethod().isMethod());
        }
    }

    private void isMethod(CharSequence isParameter) {
        if (isNameExpr.isMethod(isNameExpr)) {
            return;
        }
        if (isNameExpr.isMethod() > isIntegerConstant) {
            isNameExpr.isMethod("isStringConstant");
        }
        isNameExpr.isMethod(isNameExpr);
        isMethod();
    }

    @Override
    protected void isMethod() {
        super.isMethod();
        isNameExpr.isMethod();
        if (isNameExpr && isNameExpr) {
            isNameExpr.isMethod();
            if (isNameExpr != isNameExpr.isFieldAccessExpr) {
                isMethod();
            }
        }
        isNameExpr.isMethod().isMethod(true);
    }

    private void isMethod() {
        isNameExpr.isMethod(this, this, true).isMethod(isParameter -> {
            isNameExpr.isMethod(this, "isStringConstant" + isNameExpr);
            MyAccount isVariable = isNameExpr.isMethod().isMethod(isMethod().isMethod().isMethod());
            if (isNameExpr.isMethod()) {
                isNameExpr.isMethod().isMethod(isNameExpr);
            }
            if (isNameExpr == isNameExpr.isFieldAccessExpr) {
                if (isNameExpr)
                    isMethod(isNameExpr, isNameExpr);
                Timeline isVariable = isNameExpr.isMethod().isMethod(isNameExpr.isFieldAccessExpr, isNameExpr.isMethod(), isNameExpr.isFieldAccessExpr);
                isNameExpr.isMethod(isNameExpr, isNameExpr.this, isNameExpr, true);
                isNameExpr.isMethod();
            } else {
                if (isNameExpr.isMethod().isMethod() > isIntegerConstant) {
                    Intent isVariable = new Intent(isNameExpr.isMethod(), MySettingsActivity.class);
                    // isComment
                    isNameExpr.isMethod(isNameExpr.isFieldAccessExpr | isNameExpr.isFieldAccessExpr);
                    isMethod(isNameExpr);
                } else {
                    isNameExpr.isMethod(isNameExpr.this);
                }
            }
        });
    }

    private void isMethod(MyContext isParameter, MyAccount isParameter) {
        final Timeline isVariable = isNameExpr.isMethod().isMethod(isNameExpr.isFieldAccessExpr, isNameExpr.isMethod());
        if (!isNameExpr.isMethod())
            return;
        isNameExpr.isMethod();
        isNameExpr.isMethod(isNameExpr.isMethod(isNameExpr.isFieldAccessExpr, isNameExpr));
        isNameExpr.isMethod((isNameExpr.isMethod(isNameExpr.isFieldAccessExpr, isNameExpr.isMethod().isFieldAccessExpr, "isStringConstant")));
    }

    /**
     * isComment
     */
    private boolean isVariable = true;

    @Override
    public boolean isMethod(int isParameter, KeyEvent isParameter) {
        if (isNameExpr == isNameExpr.isFieldAccessExpr && isNameExpr.isMethod() == isIntegerConstant) {
            isMethod();
            return true;
        }
        return super.isMethod(isNameExpr, isNameExpr);
    }

    /**
     * isComment
     */
    private void isMethod() {
        // isComment
        isNameExpr.isFieldAccessExpr.isMethod();
        String isVariable = "isStringConstant";
        isNameExpr.isFieldAccessExpr = true;
        isNameExpr = isNameExpr.isFieldAccessExpr;
        if (isNameExpr.isFieldAccessExpr != null) {
            // isComment
            isNameExpr = isNameExpr.isFieldAccessExpr;
            if (isNameExpr.isFieldAccessExpr) {
                if (isNameExpr.isFieldAccessExpr.isMethod()) {
                    // isComment
                    Bundle isVariable = new Bundle();
                    isNameExpr.isMethod(isNameExpr.isFieldAccessExpr, isNameExpr.isMethod().isMethod());
                    isNameExpr.isMethod(isNameExpr.isFieldAccessExpr, isNameExpr.isFieldAccessExpr);
                    isNameExpr.isFieldAccessExpr.isMethod(isNameExpr);
                    isNameExpr += "isStringConstant" + isNameExpr.isMethod().isMethod() + "isStringConstant";
                }
            } else {
                isNameExpr.isFieldAccessExpr.isMethod(isNameExpr.isFieldAccessExpr, "isStringConstant");
            }
        }
        // isComment
        isNameExpr.isMethod();
        if (!isNameExpr) {
            isNameExpr.isMethod(this, "isStringConstant" + isNameExpr.isMethod() + "isStringConstant" + isNameExpr);
            isMethod();
        }
    }

    public static void isMethod(android.content.Context isParameter) {
        Intent isVariable;
        isNameExpr = new Intent(isNameExpr, AccountSettingsActivity.class);
        isNameExpr.isMethod(isNameExpr.isFieldAccessExpr);
        isNameExpr.isMethod(isNameExpr);
    }

    public StateOfAccountChangeProcess isMethod() {
        return isNameExpr;
    }

    /**
     * isComment
     */
    private class isClassOrIsInterface extends MyAsyncTask<Void, Void, JSONObject> {

        private ProgressDialog isVariable;

        isConstructor() {
            super(isNameExpr.isFieldAccessExpr);
        }

        @Override
        protected void isMethod() {
            isNameExpr = isNameExpr.isMethod(isNameExpr.this, isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr), isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr), // isComment
            true, // isComment
            true);
        }

        @Override
        protected JSONObject isMethod(Void... isParameter) {
            JSONObject isVariable = null;
            boolean isVariable = true;
            String isVariable = "isStringConstant";
            try {
                isNameExpr.isFieldAccessExpr.isMethod();
                if (!isNameExpr.isMethod().isMethod()) {
                    isNameExpr.isFieldAccessExpr.isMethod();
                }
                isNameExpr = isNameExpr.isMethod().isMethod();
                if (isNameExpr) {
                    isNameExpr.isFieldAccessExpr.isFieldAccessExpr.isMethod().isMethod();
                    isNameExpr.isFieldAccessExpr.isFieldAccessExpr.isMethod().isMethod();
                    isNameExpr.isFieldAccessExpr.isFieldAccessExpr.isMethod().isMethod();
                }
            } catch (ConnectionException isParameter) {
                isNameExpr = isNameExpr.isMethod();
                isNameExpr.isMethod(this, isNameExpr);
            }
            try {
                String isVariable = "isStringConstant";
                if (!isNameExpr) {
                    isNameExpr = isNameExpr.this.isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
                    if (!isNameExpr.isMethod(isNameExpr)) {
                        isNameExpr += "isStringConstant" + isNameExpr;
                    }
                    isNameExpr.isMethod(isNameExpr, isNameExpr);
                }
                isNameExpr = new JSONObject();
                isNameExpr.isMethod(isNameExpr, isNameExpr);
                isNameExpr.isMethod(isNameExpr, isNameExpr);
            } catch (JSONException isParameter) {
                isNameExpr.isMethod(this, isNameExpr);
            }
            return isNameExpr;
        }

        // isComment
        @Override
        protected void isMethod(JSONObject isParameter) {
            isNameExpr.isMethod(isNameExpr);
            boolean isVariable = true;
            if (isNameExpr != null) {
                try {
                    isNameExpr = isNameExpr.isMethod(isNameExpr);
                    String isVariable = isNameExpr.isMethod(isNameExpr);
                    if (isNameExpr) {
                        String isVariable = isNameExpr.isMethod().isMethod();
                        isNameExpr.isFieldAccessExpr = isNameExpr.isFieldAccessExpr.isMethod(isNameExpr.isMethod(), isNameExpr, isNameExpr.isFieldAccessExpr);
                        isMethod();
                        isNameExpr.isMethod(this, true, new OAuthAcquireRequestTokenTask());
                        isNameExpr = isNameExpr.isFieldAccessExpr;
                    } else {
                        isMethod(isNameExpr);
                        isNameExpr.isFieldAccessExpr.isMethod(isNameExpr.isFieldAccessExpr);
                        isMethod();
                    }
                } catch (JSONException isParameter) {
                    isNameExpr.isMethod(this, isNameExpr);
                }
            }
            isNameExpr.isMethod(this, isNameExpr.isMethod(isNameExpr));
        }
    }

    /**
     * isComment
     */
    private class isClassOrIsInterface extends MyAsyncTask<Void, Void, JSONObject> {

        private ProgressDialog isVariable;

        isConstructor() {
            super(isNameExpr.isFieldAccessExpr);
        }

        @Override
        protected void isMethod() {
            isNameExpr = isNameExpr.isMethod(isNameExpr.this, isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr), isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr), // isComment
            true, // isComment
            true);
        }

        @Override
        protected JSONObject isMethod(Void... isParameter) {
            JSONObject isVariable = null;
            boolean isVariable = true;
            String isVariable = "isStringConstant";
            String isVariable = "isStringConstant";
            try {
                MyAccount isVariable = isNameExpr.isMethod();
                isNameExpr.isMethod(this, "isStringConstant" + isNameExpr);
                String isVariable;
                if (isNameExpr.isMethod().isMethod().isMethod()) {
                    final OAuth20Service isVariable = isNameExpr.isMethod().isMethod().isMethod(true);
                    isNameExpr = isNameExpr.isMethod();
                } else {
                    OAuthConsumer isVariable = isNameExpr.isMethod().isMethod().isMethod();
                    // isComment
                    // isComment
                    // isComment
                    // isComment
                    // isComment
                    isNameExpr = isNameExpr.isMethod().isMethod().isMethod().isMethod(isNameExpr, isNameExpr.isFieldAccessExpr.isMethod());
                    isNameExpr.isMethod(isNameExpr.isMethod(), isNameExpr.isMethod());
                }
                // isComment
                // isComment
                isNameExpr.isFieldAccessExpr = true;
                // isComment
                Intent isVariable = new Intent(isNameExpr.this, AccountSettingsWebActivity.class);
                isNameExpr.isMethod(isNameExpr.isFieldAccessExpr, isNameExpr);
                isNameExpr.this.isMethod(isNameExpr);
                isNameExpr = true;
            } catch (OAuthMessageSignerException | OAuthNotAuthorizedException | OAuthExpectationFailedException | OAuthCommunicationException | ConnectionException isParameter) {
                isNameExpr = isNameExpr.isMethod();
                isNameExpr.isMethod(this, isNameExpr);
            }
            try {
                if (!isNameExpr) {
                    isNameExpr = isNameExpr.this.isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
                    if (isNameExpr != null && isNameExpr.isMethod() > isIntegerConstant) {
                        isNameExpr += "isStringConstant" + isNameExpr;
                    }
                    isNameExpr.isMethod(isNameExpr, isNameExpr);
                    isNameExpr.isFieldAccessExpr.isMethod();
                }
                isNameExpr = new JSONObject();
                isNameExpr.isMethod(isNameExpr, isNameExpr);
                isNameExpr.isMethod(isNameExpr, isNameExpr);
            } catch (JSONException isParameter) {
                isNameExpr.isMethod(this, isNameExpr);
            }
            return isNameExpr;
        }

        // isComment
        @Override
        protected void isMethod(JSONObject isParameter) {
            isNameExpr.isMethod(isNameExpr);
            boolean isVariable = true;
            if (isNameExpr != null) {
                try {
                    isNameExpr = isNameExpr.isMethod(isNameExpr);
                    String isVariable = isNameExpr.isMethod(isNameExpr);
                    if (isNameExpr) {
                        // isComment
                        // isComment
                        // isComment
                        isNameExpr.this.isMethod();
                    } else {
                        isMethod(isNameExpr);
                        isNameExpr.isFieldAccessExpr.isMethod(isNameExpr.isFieldAccessExpr);
                        isMethod();
                    }
                } catch (JSONException isParameter) {
                    isNameExpr.isMethod(this, isNameExpr);
                }
            }
            isNameExpr.isMethod(this, isNameExpr.isMethod(isNameExpr));
        }
    }

    /**
     * isComment
     */
    private class isClassOrIsInterface extends MyAsyncTask<Uri, Void, JSONObject> {

        private ProgressDialog isVariable;

        isConstructor() {
            super(isNameExpr.isFieldAccessExpr);
        }

        @Override
        protected void isMethod() {
            isNameExpr = isNameExpr.isMethod(isNameExpr.this, isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr), isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr), // isComment
            true, // isComment
            true);
        }

        @Override
        protected JSONObject isMethod(Uri... isParameter) {
            String isVariable = "isStringConstant";
            String isVariable = "isStringConstant";
            String isVariable = "isStringConstant";
            if (isNameExpr.isMethod().isMethod() == null) {
                isNameExpr = "isStringConstant";
                isNameExpr.isMethod(this, isNameExpr);
            } else {
                // isComment
                Uri isVariable = isNameExpr[isIntegerConstant];
                if (isNameExpr != null && isNameExpr.isFieldAccessExpr.isMethod().isMethod(isNameExpr.isMethod())) {
                    isNameExpr.isFieldAccessExpr.isMethod(isNameExpr.isFieldAccessExpr);
                    try {
                        if (isNameExpr.isMethod().isMethod().isMethod()) {
                            String isVariable = isNameExpr.isMethod("isStringConstant");
                            isNameExpr.isMethod(this, "isStringConstant" + isNameExpr);
                            final OAuth20Service isVariable = isNameExpr.isMethod().isMethod().isMethod(true);
                            final OAuth2AccessToken isVariable = isNameExpr.isMethod(isNameExpr);
                            isNameExpr = isNameExpr.isMethod();
                            isNameExpr = isNameExpr.isMethod();
                        } else {
                            String isVariable = isNameExpr.isMethod();
                            String isVariable = isNameExpr.isMethod();
                            // isComment
                            isNameExpr.isMethod(null, null);
                            OAuthConsumer isVariable = isNameExpr.isMethod().isMethod().isMethod();
                            if (!(isNameExpr == null || isNameExpr == null)) {
                                isNameExpr.isMethod(isNameExpr, isNameExpr);
                            }
                            String isVariable = isNameExpr.isMethod(isNameExpr.isFieldAccessExpr);
                            String isVariable = isNameExpr.isMethod(isNameExpr.isFieldAccessExpr);
                            /*isComment*/
                            if (isNameExpr != null || isNameExpr.isMethod() != null) {
                                isNameExpr.isMethod().isMethod().isMethod().isMethod(isNameExpr, isNameExpr);
                                // isComment
                                isNameExpr = isNameExpr.isMethod();
                                isNameExpr = isNameExpr.isMethod();
                            }
                        }
                    } catch (Exception isParameter) {
                        isNameExpr = isNameExpr.isMethod();
                        isNameExpr.isMethod(this, isNameExpr);
                    } finally {
                        isNameExpr.isFieldAccessExpr.isMethod(isNameExpr, isNameExpr);
                        isNameExpr.isMethod(this, "isStringConstant" + isNameExpr.isMethod().isMethod() + "isStringConstant" + isNameExpr + "isStringConstant" + isNameExpr);
                    }
                }
            }
            JSONObject isVariable = null;
            try {
                isNameExpr = new JSONObject();
                isNameExpr.isMethod(isNameExpr, !isNameExpr.isMethod(isNameExpr) && !isNameExpr.isMethod(isNameExpr));
                isNameExpr.isMethod(isNameExpr, isNameExpr);
            } catch (JSONException isParameter) {
                isNameExpr.isMethod(this, isNameExpr);
            }
            return isNameExpr;
        }

        // isComment
        @Override
        protected void isMethod(JSONObject isParameter) {
            isNameExpr.isMethod(isNameExpr);
            boolean isVariable = true;
            if (isNameExpr != null) {
                try {
                    isNameExpr = isNameExpr.isMethod(isNameExpr);
                    String isVariable = isNameExpr.isMethod(isNameExpr);
                    if (isNameExpr) {
                        // isComment
                        // isComment
                        isNameExpr.isMethod(this, true, new VerifyCredentialsTask());
                    } else {
                        String isVariable = isNameExpr.this.isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
                        if (!isNameExpr.isMethod(isNameExpr)) {
                            isNameExpr += "isStringConstant" + isNameExpr;
                            isNameExpr.isMethod(isNameExpr, isNameExpr);
                        }
                        isMethod(isNameExpr);
                        isNameExpr.isFieldAccessExpr.isMethod(isNameExpr.isFieldAccessExpr);
                        isMethod();
                    }
                } catch (JSONException isParameter) {
                    isNameExpr.isMethod(this, isNameExpr);
                }
            }
            isNameExpr.isMethod(this, isNameExpr.isMethod(isNameExpr));
        }
    }

    /**
     * isComment
     */
    private class isClassOrIsInterface extends MyAsyncTask<Void, Void, TaskResult> {

        private ProgressDialog isVariable;

        private volatile boolean isVariable = true;

        isConstructor() {
            super(isNameExpr.isFieldAccessExpr);
        }

        @Override
        protected void isMethod() {
            isNameExpr = isNameExpr.isFieldAccessExpr;
            isNameExpr = isNameExpr.isMethod(isNameExpr.this, isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr), isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr), // isComment
            true, // isComment
            true);
            synchronized (isNameExpr.this) {
                if (isNameExpr) {
                    isNameExpr = true;
                } else {
                    isNameExpr = true;
                }
            }
        }

        @Override
        protected TaskResult isMethod(Void... isParameter) {
            if (isNameExpr)
                return new TaskResult(isNameExpr.isFieldAccessExpr);
            ResultStatus isVariable = isNameExpr.isFieldAccessExpr;
            String isVariable = "isStringConstant";
            try {
                isNameExpr.isFieldAccessExpr.isMethod();
                isNameExpr.isFieldAccessExpr.isMethod();
                final MyAccount isVariable = isNameExpr.isFieldAccessExpr.isMethod();
                if (isNameExpr.isMethod()) {
                    isNameExpr.isMethod();
                    MyContext isVariable = isNameExpr.isMethod(isNameExpr.isMethod().isMethod(), isNameExpr.this);
                    isNameExpr = isNameExpr.isFieldAccessExpr;
                    final Timeline isVariable = isNameExpr.isMethod().isMethod(isNameExpr.isFieldAccessExpr, isNameExpr.isMethod());
                    if (isNameExpr.isMethod()) {
                        isNameExpr = true;
                        isNameExpr = isNameExpr.isFieldAccessExpr;
                    }
                }
            } catch (ConnectionException isParameter) {
                switch(isNameExpr.isMethod()) {
                    case isNameExpr:
                        isNameExpr = isNameExpr.isFieldAccessExpr;
                        break;
                    case isNameExpr:
                        isNameExpr = isNameExpr.isFieldAccessExpr;
                        break;
                    default:
                        isNameExpr = isNameExpr.isFieldAccessExpr;
                        break;
                }
                isNameExpr = isNameExpr.isMethod();
                isNameExpr.isMethod(this, isNameExpr);
            }
            return new TaskResult(isNameExpr, isNameExpr);
        }

        /**
         * isComment
         */
        @Override
        protected void isMethod(TaskResult isParameter) {
            isNameExpr.isMethod(isNameExpr);
            TaskResult isVariable = isNameExpr == null ? new TaskResult(isNameExpr.isFieldAccessExpr) : isNameExpr;
            CharSequence isVariable = "isStringConstant";
            switch(isNameExpr.isFieldAccessExpr) {
                case isNameExpr:
                    isNameExpr.isMethod(isNameExpr.this, isNameExpr.isFieldAccessExpr.isFieldAccessExpr, isNameExpr.isFieldAccessExpr).isMethod();
                    break;
                case isNameExpr:
                    isNameExpr = isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
                    break;
                case isNameExpr:
                    isNameExpr = isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
                    break;
                case isNameExpr:
                    isNameExpr = isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr) + "isStringConstant" + isNameExpr.isFieldAccessExpr;
                    isNameExpr.isMethod(this, isNameExpr.isMethod());
                    break;
                default:
                    break;
            }
            if (isNameExpr == isNameExpr.isFieldAccessExpr) {
                if (isMethod()) {
                    isMethod();
                } else {
                    isMethod();
                    isMethod();
                }
                return;
            }
            if (!isNameExpr) {
                StateOfAccountChangeProcess isVariable = isNameExpr.this.isFieldAccessExpr;
                // isComment
                // isComment
                isNameExpr.isFieldAccessExpr = isNameExpr.isMethod();
                if (isNameExpr.isMethod()) {
                    isNameExpr.isFieldAccessExpr = true;
                    if (isNameExpr.isMethod().isMethod(isNameExpr.isFieldAccessExpr) == isIntegerConstant) {
                        isNameExpr.isMethod(isNameExpr.isFieldAccessExpr);
                    }
                }
                isNameExpr = true;
            }
            isMethod();
            isMethod(isNameExpr);
        }
    }
}
