ast_results/andstatus_andstatus/app/src/androidTest/java/org/andstatus/app/account/AccountSettingsActivityTest.java
// isComment
package org.andstatus.app.account;

import android.app.Activity;
import android.app.Instrumentation.ActivityMonitor;
import android.content.Intent;
import android.widget.Button;
import android.widget.TextView;
import org.andstatus.app.IntentExtra;
import org.andstatus.app.R;
import org.andstatus.app.context.ActivityTest;
import org.andstatus.app.context.MyContextHolder;
import org.andstatus.app.context.MySettingsActivity;
import org.andstatus.app.context.TestSuite;
import org.andstatus.app.data.DbUtils;
import org.andstatus.app.origin.OriginType;
import org.andstatus.app.origin.PersistentOriginList;
import org.andstatus.app.service.MyServiceManager;
import org.andstatus.app.util.MyLog;
import org.junit.Test;
import static org.junit.Assert.assertEquals;
import static org.junit.Assert.assertNotNull;
import static org.junit.Assert.assertTrue;
import static org.junit.Assert.fail;

/**
 * isComment
 */
public class isClassOrIsInterface extends ActivityTest<AccountSettingsActivity> {

    private MyAccount isVariable = null;

    @Override
    protected Class<AccountSettingsActivity> isMethod() {
        return AccountSettingsActivity.class;
    }

    @Override
    protected Intent isMethod() {
        isNameExpr.isMethod(this);
        isNameExpr = isNameExpr.isMethod().isMethod().isMethod();
        if (isNameExpr.isMethod())
            isMethod("isStringConstant");
        return new Intent().isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr, isNameExpr.isMethod());
    }

    @Test
    public void isMethod() throws InterruptedException {
        final String isVariable = "isStringConstant";
        Button isVariable = (Button) isMethod().isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
        isMethod(isNameExpr != null);
        isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
        isMethod(isNameExpr.isFieldAccessExpr.isFieldAccessExpr);
        isNameExpr.isMethod(isNameExpr, isIntegerConstant);
        isMethod("isStringConstant", true, isNameExpr.isMethod());
        isMethod();
        isNameExpr.isMethod(isNameExpr, isIntegerConstant);
        isMethod().isMethod();
        isNameExpr.isMethod(isNameExpr, isIntegerConstant);
        isNameExpr.isMethod(isMethod().isMethod());
        isNameExpr.isMethod(isNameExpr, isIntegerConstant);
    }

    private void isMethod(int isParameter) {
        TextView isVariable = (TextView) isMethod().isMethod(isNameExpr);
        isMethod(isNameExpr != null);
        isMethod("isStringConstant", isNameExpr.isMethod(), isNameExpr.isMethod().isMethod());
    }

    private void isMethod() throws InterruptedException {
        final String isVariable = "isStringConstant";
        ActivityMonitor isVariable = isMethod().isMethod(PersistentOriginList.class.isMethod(), null, true);
        Runnable isVariable = new Runnable() {

            @Override
            public void isMethod() {
                isNameExpr.isMethod(this, isNameExpr + "isStringConstant");
                isMethod().isMethod(isNameExpr.isFieldAccessExpr);
            }
        };
        isNameExpr.isMethod(this, isNameExpr + "isStringConstant");
        isMethod().isMethod(isNameExpr);
        Activity isVariable = isMethod().isMethod(isNameExpr, isIntegerConstant);
        isNameExpr.isMethod(this, isNameExpr + "isStringConstant" + isNameExpr);
        isMethod("isStringConstant", isNameExpr);
        isNameExpr.isMethod(isNameExpr, isIntegerConstant);
        isNameExpr.isMethod(isNameExpr, isIntegerConstant);
        isNameExpr.isMethod();
    }
}
